Move EditVote to src
